+++ packages/frontend/app/(application)/accounts/add/page.tsx
@@ -1,19 +1,14 @@
 import { Metadata } from 'next';
 import { FC } from 'react';
 
-import { AccountAddContainer } from '$container/accounts/account.add.container';
-import { Layout } from '$layouts/Layout';
+import { AccountAddContainer } from '$container/accounts/AccountAddContainer';
 
 export const metadata: Metadata = {
   title: 'Add Account',
 };
 
 const AddAccountPage: FC = () => {
-  return (
-    <Layout title="Add Account">
-      <AccountAddContainer />
-    </Layout>
-  );
+  return <AccountAddContainer />;
 };
 
 export default AddAccountPage;
