+++ packages/frontend/app/(application)/accounts/page.tsx
@@ -1,19 +1,14 @@
 import { Metadata } from 'next';
 import { FC } from 'react';
 
-import { AccountListingContainer } from '$container/accounts/account.listing.container';
-import { Layout } from '$layouts/Layout';
+import { AccountListingContainer } from '$container/accounts/AccountListingContainer';
 
 export const metadata: Metadata = {
   title: 'Accounts',
 };
 
 const AccountPage: FC = () => {
-  return (
-    <Layout title="Accounts">
-      <AccountListingContainer />
-    </Layout>
-  );
+  return <AccountListingContainer />;
 };
 
 export default AccountPage;
